+++ classifiers/OC1_source/impurity_measures.c
@@ -31,7 +31,7 @@
 /*                      number of  classes (categories) and the */
 /*                      counts of points of each class on the   */
 /*                      left and right of the hyperplane.       */
-/*                      All measures return a nonnegative float */
+/*                      All measures return a nonnegative double */
 /*                      impurity value, where the lower the     */
 /*                      impurity, the better the hyperplane.    */
 /*                      All measures are called from the module */
@@ -58,7 +58,7 @@ int largest_element();
 /*		except the class with the highest count.)		*/
 /* Calls modules :	largest_element (compute_impurity.c)		*/
 /************************************************************************/
-float maxminority()
+double maxminority()
 {
   int i, j, lminor = 0, rminor = 0;
 
@@ -75,9 +75,9 @@ float maxminority()
         rminor += right_count[j];
 
   if (lminor > rminor)
-    return ((float)lminor);
+    return ((double)lminor);
   else
-    return ((float)rminor);
+    return ((double)rminor);
 }
 
 /************************************************************************/
@@ -92,7 +92,7 @@ float maxminority()
 /*		except the class with the highest count.)		*/
 /* Calls modules :	largest_element (compute_impurity.c)		*/
 /************************************************************************/
-float summinority()
+double summinority()
 {
   int i, j, lminor = 0, rminor = 0;
 
@@ -108,7 +108,7 @@ float summinority()
       if (i != j)
         rminor += right_count[j];
 
-  return ((float)(lminor + rminor));
+  return ((double)(lminor + rminor));
 }
 
 /************************************************************************/
@@ -123,9 +123,9 @@ float summinority()
 /*		points on that side of the hyperplane, and		*/
 /*		avg = (x1+x2+..+xk)/k.					*/
 /************************************************************************/
-float variance()
+double variance()
 {
-  float lavg = 0, ravg = 0, lerror = 0, rerror = 0;
+  double lavg = 0, ravg = 0, lerror = 0, rerror = 0;
   int i, lsum1 = 0, rsum1 = 0, lsum2 = 0, rsum2 = 0;
   int *temp1 = NULL, *temp2 = NULL;
   int var_compare();
@@ -155,9 +155,9 @@ float variance()
   }
 
   if (lsum1 != 0)
-    lavg = (float)lsum2 / lsum1;
+    lavg = (double)lsum2 / lsum1;
   if (rsum1 != 0)
-    ravg = (float)rsum2 / rsum1;
+    ravg = (double)rsum2 / rsum1;
 
   for (i = 1; i <= no_of_categories; i++)
   {
@@ -205,12 +205,12 @@ int var_compare(p1, p2) int *p1, *p2;
 /* Remarks : Much less efficient to compute than the minority measures. */
 /*           But often works much better.                               */
 /************************************************************************/
-float info_gain()  // TAG: infogain
+double info_gain()  // TAG: infogain
 {
-  float presplit_info = 0, postsplit_info = 0, left_info = 0, right_info = 0;
-  float ratio, infogain;
+  double presplit_info = 0, postsplit_info = 0, left_info = 0, right_info = 0;
+  double ratio, infogain;
   int i, total_count = 0, total_left_count = 0, total_right_count = 0;
-  float mylog2();
+  double mylog2();
 
   for (i = 1; i <= no_of_categories; i++)
   {
@@ -222,7 +222,7 @@ float info_gain()  // TAG: infogain
   if (total_count)
     for (i = 1; i <= no_of_categories; i++)
     {
-      ratio = (float)(left_count[i] + right_count[i]) / total_count;
+      ratio = (double)(left_count[i] + right_count[i]) / total_count;
       if (ratio)
         presplit_info += -1.0 * ratio * mylog2(ratio);
     }
@@ -231,7 +231,7 @@ float info_gain()  // TAG: infogain
   {
     for (i = 1; i <= no_of_categories; i++)
     {
-      ratio = (float)left_count[i] / total_left_count;
+      ratio = (double)left_count[i] / total_left_count;
       if (ratio)
         left_info += -1.0 * ratio * mylog2(ratio);
     }
@@ -242,7 +242,7 @@ float info_gain()  // TAG: infogain
   {
     for (i = 1; i <= no_of_categories; i++)
     {
-      ratio = (float)right_count[i] / total_right_count;
+      ratio = (double)right_count[i] / total_right_count;
       if (ratio)
         right_info += -1.0 * ratio * mylog2(ratio);
     }
@@ -279,10 +279,10 @@ float info_gain()  // TAG: infogain
 /* Remarks : Efficient to compute - No log computations.                */
 /*           Performs quite well.                                       */
 /************************************************************************/
-float gini_index()
+double gini_index()
 {
   int total_left_count = 0, total_right_count = 0;
-  float temp, gini_left = 0, gini_right = 0, gini_value;
+  double temp, gini_left = 0, gini_right = 0, gini_value;
   int i, j;
 
   for (i = 1; i <= no_of_categories; i++)
@@ -323,10 +323,10 @@ float gini_index()
 /*			The twoing measure is described in detail  	*/
 /*			by Leo Breiman et al in their CART book (1984).	*/
 /************************************************************************/
-float twoing()
+double twoing()
 {
-  float total_left_count = 0, total_right_count = 0, total_count;
-  float goodness = 0, temp, twoing_val;
+  double total_left_count = 0, total_right_count = 0, total_count;
+  double goodness = 0, temp, twoing_val;
   int i;
 
   for (i = 1; i <= no_of_categories; i++)
